requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.26.4
scikit-learn==1.5.1
python-docx==0.8.11
PyPDF2==3.0.1
pickle-mixin==1.0.2
